META-3040 Addressed review comments
