+++ graphdb/janus/src/main/java/org/apache/atlas/repository/graphdb/janus/AtlasJanusGraph.java
@@ -94,6 +94,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import static org.apache.atlas.AtlasErrorCode.INDEX_SEARCH_FAILED;
 import static org.apache.atlas.AtlasErrorCode.RELATIONSHIP_CREATE_INVALID_PARAMS;
 import static org.apache.atlas.ESAliasRequestBuilder.ESAliasAction.REMOVE;
 import static org.apache.atlas.repository.Constants.*;
@@ -315,17 +316,19 @@ public class AtlasJanusGraph implements AtlasGraph<AtlasJanusVertex, AtlasJanusE
         return new AtlasElasticsearchQuery(this, elasticsearchClient, INDEX_PREFIX + indexName, sourceBuilder);
     }
 
-    public AtlasIndexQuery<AtlasJanusVertex, AtlasJanusEdge> elasticsearchQuery(String indexName, SearchParams searchParams) {
+    public AtlasIndexQuery<AtlasJanusVertex, AtlasJanusEdge> elasticsearchQuery(String indexName, SearchParams searchParams)throws AtlasBaseException {
         if (restClient == null) {
             LOG.error("restClient is not initiated, failed to run query on ES");
+            throw new AtlasBaseException(INDEX_SEARCH_FAILED, "restClient is not initiated");
         }
         return new AtlasElasticsearchQuery(this, restClient, INDEX_PREFIX + indexName, searchParams);
     }
 
     @Override
-    public AtlasElasticsearchQuery elasticsearchQuery(String indexName) {
+    public AtlasElasticsearchQuery elasticsearchQuery(String indexName) throws AtlasBaseException {
         if (restClient == null) {
             LOG.error("restClient is not initiated, failed to run query on ES");
+            throw new AtlasBaseException(INDEX_SEARCH_FAILED, "restClient is not initiated");
         }
         return new AtlasElasticsearchQuery(this, indexName, restClient);
     }
