+++ repository/src/main/java/org/apache/atlas/accesscontrol/purpose/AtlasPurposeService.java
@@ -30,25 +30,16 @@ import org.apache.atlas.utils.AtlasPerfMetrics;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.ranger.plugin.model.RangerPolicy;
-import org.apache.ranger.plugin.model.RangerPolicy.RangerDataMaskPolicyItem;
-import org.apache.ranger.plugin.model.RangerPolicy.RangerPolicyItemAccess;
-import org.apache.ranger.plugin.model.RangerPolicy.RangerPolicyItemDataMaskInfo;
 import org.apache.ranger.plugin.model.RangerPolicyResourceSignature;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static org.apache.atlas.AtlasErrorCode.BAD_REQUEST;
 import static org.apache.atlas.AtlasErrorCode.OPERATION_NOT_SUPPORTED;
-import static org.apache.atlas.AtlasErrorCode.PURPOSE_ALREADY_EXISTS;
 import static org.apache.atlas.accesscontrol.AccessControlUtil.fetchRangerPoliciesByLabel;
 import static org.apache.atlas.accesscontrol.AccessControlUtil.getESAliasName;
 import static org.apache.atlas.accesscontrol.persona.AtlasPersonaUtil.getIsEnabled;
@@ -86,19 +77,23 @@ public class AtlasPurposeService {
 
         context.setCreateNewPurpose(true);
 
-        validateUniquenessByName(graph, getName(entityWithExtInfo.getEntity()), PURPOSE_ENTITY_TYPE);
-        validateUniquenessByTags(graph, getTags(entityWithExtInfo.getEntity()), PURPOSE_ENTITY_TYPE);
+        try {
+            validateUniquenessByName(graph, getName(entityWithExtInfo.getEntity()), PURPOSE_ENTITY_TYPE);
+            validateUniquenessByTags(graph, getTags(entityWithExtInfo.getEntity()), PURPOSE_ENTITY_TYPE);
 
-        String tenantId = getTenantId(context.getPurpose());
-        if (StringUtils.isEmpty(tenantId)) {
-            tenantId = "tenant";
+            //unique qualifiedName for Purpose
+            String tenantId = getTenantId(context.getPurpose());
+            if (StringUtils.isEmpty(tenantId)) {
+                tenantId = "tenant";
+            }
+            entityWithExtInfo.getEntity().setAttribute(QUALIFIED_NAME, String.format(POLICY_QN_FORMAT, tenantId, getUUID()));
+            entityWithExtInfo.getEntity().setAttribute("enabled", true);
+
+            aliasStore.createAlias(context);
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
         }
-        entityWithExtInfo.getEntity().setAttribute(QUALIFIED_NAME, String.format(POLICY_QN_FORMAT, tenantId, getUUID()));
-        entityWithExtInfo.getEntity().setAttribute("enabled", true);
 
-        aliasStore.createAlias(context);
-
-        RequestContext.get().endMetricRecord(metricRecorder);
         return ret;
     }
 
@@ -112,41 +107,49 @@ public class AtlasPurposeService {
 
         purpose.setAttribute(QUALIFIED_NAME, getQualifiedName(existingPurposeEntity));
 
-        if (!AtlasEntity.Status.ACTIVE.equals(existingPurposeEntity.getStatus())) {
-            throw new AtlasBaseException(OPERATION_NOT_SUPPORTED, "Purpose not Active");
-        }
-
-        if (getIsEnabled(existingPurposeEntity) != getIsEnabled(purpose)) {
-            if (getIsEnabled(purpose)) {
-                enablePurpose(existingPurposeWithExtInfo);
-            } else {
-                disablePurpose(existingPurposeWithExtInfo);
+        try {
+            if (!AtlasEntity.Status.ACTIVE.equals(existingPurposeEntity.getStatus())) {
+                throw new AtlasBaseException(OPERATION_NOT_SUPPORTED, "Purpose not Active");
             }
-        }
 
-        if (!getName(purpose).equals(getName(existingPurposeEntity))) {
-            validateUniquenessByName(graph, getName(purpose), PURPOSE_ENTITY_TYPE);
-        }
+            if (getIsEnabled(existingPurposeEntity) != getIsEnabled(purpose)) {
+                if (getIsEnabled(purpose)) {
+                    enablePurpose(existingPurposeWithExtInfo);
+                } else {
+                    disablePurpose(existingPurposeWithExtInfo);
+                }
+            }
 
-        if (!CollectionUtils.isEqualCollection(getTags(purpose), getTags(existingPurposeEntity))) {
-            needPolicyUpdate = true;
-            validateUniquenessByTags(graph, getTags(purpose), PURPOSE_ENTITY_TYPE);
-        }
+            //check name update
+            // if yes: check naw name for uniqueness
+            if (!getName(purpose).equals(getName(existingPurposeEntity))) {
+                validateUniquenessByName(graph, getName(purpose), PURPOSE_ENTITY_TYPE);
+            }
 
-        if (needPolicyUpdate && CollectionUtils.isNotEmpty(getPolicies(context.getPurposeExtInfo()))) {
-            updatePurposePoliciesTag(context, existingPurposeEntity);
-            aliasStore.updateAlias(context);
-        }
+            //check tags update
+            // if yes: check tags for uniqueness
+            if (!CollectionUtils.isEqualCollection(getTags(purpose), getTags(existingPurposeEntity))) {
+                needPolicyUpdate = true;
+                validateUniquenessByTags(graph, getTags(purpose), PURPOSE_ENTITY_TYPE);
+            }
 
-        RequestContext.get().endMetricRecord(metricRecorder);
+            if (needPolicyUpdate && CollectionUtils.isNotEmpty(getPolicies(context.getPurposeExtInfo()))) {
+                updatePurposePoliciesTag(context, existingPurposeEntity);
+                aliasStore.updateAlias(context);
+            }
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
+        }
     }
 
     private void disablePurpose(AtlasEntityWithExtInfo existingPurposeWithExtInfo) throws AtlasBaseException {
         AtlasPerfMetrics.MetricRecorder metricRecorder = RequestContext.get().startMetricRecord("disablePurpose");
 
-        cleanRangerPolicies(existingPurposeWithExtInfo.getEntity());
-
-        RequestContext.get().endMetricRecord(metricRecorder);
+        try {
+            cleanRangerPolicies(existingPurposeWithExtInfo.getEntity());
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
+        }
     }
 
     private void enablePurpose(AtlasEntityWithExtInfo existingPurposeWithExtInfo) throws AtlasBaseException {
@@ -155,10 +158,12 @@ public class AtlasPurposeService {
         PurposeContext context = new PurposeContext();
         context.setPurposeExtInfo(existingPurposeWithExtInfo);
 
-        List<RangerPolicy> provisionalRangerPolicies = purposePolicyToRangerPolicy(context);
-        updatePurposePolicies(context, provisionalRangerPolicies);
-
-        RequestContext.get().endMetricRecord(metricRecorder);
+        try {
+            List<RangerPolicy> provisionalRangerPolicies = PurposeServiceHelper.purposePolicyToRangerPolicy(context);
+            updatePurposePolicies(context, provisionalRangerPolicies);
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
+        }
     }
 
     public void deletePurpose(AtlasEntityWithExtInfo personaExtInfo) throws AtlasBaseException {
@@ -169,16 +174,18 @@ public class AtlasPurposeService {
             throw new AtlasBaseException(BAD_REQUEST, "Please provide entity of type " + PURPOSE_ENTITY_TYPE);
         }
 
-        if(!purpose.getStatus().equals(AtlasEntity.Status.ACTIVE)) {
-            LOG.info("Purpose with guid {} is already deleted/purged", purpose.getGuid());
-            return;
+        try {
+            if(!purpose.getStatus().equals(AtlasEntity.Status.ACTIVE)) {
+                LOG.info("Purpose with guid {} is already deleted/purged", purpose.getGuid());
+                return;
+            }
+
+            cleanRangerPolicies(purpose);
+
+            aliasStore.deleteAlias(getESAliasName(purpose));
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
         }
-
-        cleanRangerPolicies(purpose);
-
-        aliasStore.deleteAlias(getESAliasName(purpose));
-
-        RequestContext.get().endMetricRecord(metricRecorder);
     }
 
     private void cleanRangerPolicies(AtlasEntity purpose) throws AtlasBaseException {
@@ -204,26 +211,28 @@ public class AtlasPurposeService {
 
         AtlasEntity purposePolicy = context.getPurposePolicy();
 
-        if(!POLICY_ENTITY_TYPE.equals(purposePolicy.getTypeName())) {
-            throw new AtlasBaseException(BAD_REQUEST, "Please provide entity of type " + POLICY_ENTITY_TYPE);
+        try {
+            if(!POLICY_ENTITY_TYPE.equals(purposePolicy.getTypeName())) {
+                throw new AtlasBaseException(BAD_REQUEST, "Please provide entity of type " + POLICY_ENTITY_TYPE);
+            }
+
+            if(!purposePolicy.getStatus().equals(AtlasEntity.Status.ACTIVE)) {
+                LOG.info("Purpose policy with guid {} is already deleted/purged", purposePolicy.getGuid());
+                return;
+            }
+
+            List<RangerPolicy> provisionalRangerPolicies = PurposeServiceHelper.purposePolicyToRangerPolicy(context);
+            updatePurposePolicies(context, provisionalRangerPolicies);
+
+            processDeletePolicy(context, provisionalRangerPolicies);
+
+            List<String> actions = getActions(purposePolicy);
+            if (actions.contains(ACCESS_ENTITY_READ)) {
+                aliasStore.updateAlias(context);
+            }
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
         }
-
-        if(!purposePolicy.getStatus().equals(AtlasEntity.Status.ACTIVE)) {
-            LOG.info("Purpose policy with guid {} is already deleted/purged", purposePolicy.getGuid());
-            return;
-        }
-
-        List<RangerPolicy> provisionalRangerPolicies = purposePolicyToRangerPolicy(context);
-        updatePurposePolicies(context, provisionalRangerPolicies);
-
-        processDeletePolicy(context, provisionalRangerPolicies);
-
-        List<String> actions = getActions(purposePolicy);
-        if (actions.contains(ACCESS_ENTITY_READ)) {
-            aliasStore.updateAlias(context);
-        }
-
-        RequestContext.get().endMetricRecord(metricRecorder);
     }
 
     private void processDeletePolicy(PurposeContext context, List<RangerPolicy> provisionalRangerPolicies) throws AtlasBaseException {
@@ -254,16 +263,19 @@ public class AtlasPurposeService {
         context.setAllowPolicy(getIsAllow(purposePolicy));
         context.setAllowPolicyUpdate();
 
-        validatePurposePolicy(context);
+        try {
+            PurposeServiceHelper.validatePurposePolicy(context);
 
-        purposePolicy.setAttribute(QUALIFIED_NAME, String.format(POLICY_QN_FORMAT, getQualifiedName(context.getPurpose()), getUUID()));
+            purposePolicy.setAttribute(QUALIFIED_NAME, String.format(POLICY_QN_FORMAT, getQualifiedName(context.getPurpose()), getUUID()));
 
-        List<RangerPolicy> provisionalRangerPolicies = purposePolicyToRangerPolicy(context);
-        updatePurposePolicies(context, provisionalRangerPolicies);
+            List<RangerPolicy> provisionalRangerPolicies = PurposeServiceHelper.purposePolicyToRangerPolicy(context);
 
-        aliasStore.updateAlias(context);
+            updatePurposePolicies(context, provisionalRangerPolicies);
 
-        RequestContext.get().endMetricRecord(metricRecorder);
+            aliasStore.updateAlias(context);
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
+        }
     }
 
     public void updatePurposePolicy(PurposeContext context) throws AtlasBaseException  {
@@ -274,71 +286,75 @@ public class AtlasPurposeService {
         context.setAllowPolicy(getIsAllow(purposePolicy));
         context.setAllowPolicyUpdate();
 
-        validatePurposePolicy(context);
+        try {
+            PurposeServiceHelper.validatePurposePolicy(context);
 
-        List<RangerPolicy> provisionalRangerPolicies = purposePolicyToRangerPolicy(context);
-        updatePurposePolicies(context, provisionalRangerPolicies);
+            List<RangerPolicy> provisionalRangerPolicies = PurposeServiceHelper.purposePolicyToRangerPolicy(context);
+            updatePurposePolicies(context, provisionalRangerPolicies);
 
-        aliasStore.updateAlias(context);
-
-        RequestContext.get().endMetricRecord(metricRecorder);
-    }
-
-    private void validatePurposePolicy(PurposeContext context) throws AtlasBaseException {
-        validatePurposePolicyRequest(context);
-        verifyUniqueNameForPurposePolicy(context);
+            aliasStore.updateAlias(context);
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
+        }
     }
 
     private void updatePurposePolicies(PurposeContext context, List<RangerPolicy> provisionalPolicies) throws AtlasBaseException {
         AtlasPerfMetrics.MetricRecorder metricRecorder = RequestContext.get().startMetricRecord("updatePurposePolicies");
+        RangerPolicy ret;
 
-        if (context.getExistingPurposePolicy() != null && !AtlasEntity.Status.ACTIVE.equals(context.getExistingPurposePolicy().getStatus())) {
-            throw new AtlasBaseException(OPERATION_NOT_SUPPORTED, "Entity not Active");
-        }
+        try {
+            if (context.getExistingPurposePolicy() != null && !AtlasEntity.Status.ACTIVE.equals(context.getExistingPurposePolicy().getStatus())) {
+                throw new AtlasBaseException(OPERATION_NOT_SUPPORTED, "Entity not Active");
+            }
 
-        for (RangerPolicy provisionalPolicy : provisionalPolicies) {
-            List<RangerPolicy> rangerPolicies = fetchRangerPoliciesByLabel(atlasRangerService,
-                    "tag",
-                    String.valueOf(provisionalPolicy.getPolicyType()),
-                    getPurposeLabel(context.getPurpose().getGuid()));
-
-            RangerPolicy rangerPolicy = null;
-            if (CollectionUtils.isEmpty(rangerPolicies)) {
-                rangerPolicy = fetchRangerPolicyByResources(atlasRangerService,
+            for (RangerPolicy provisionalPolicy : provisionalPolicies) {
+                List<RangerPolicy> rangerPolicies = fetchRangerPoliciesByLabel(atlasRangerService,
                         "tag",
                         String.valueOf(provisionalPolicy.getPolicyType()),
-                        provisionalPolicy);
+                        getPurposeLabel(context.getPurpose().getGuid()));
 
-            } else if (rangerPolicies.size() == 1) {
-                rangerPolicy = rangerPolicies.get(0);
-            } else {
-                String resourcesSignature = new RangerPolicyResourceSignature(provisionalPolicy).getSignature();
+                RangerPolicy rangerPolicy = null;
+                if (CollectionUtils.isEmpty(rangerPolicies)) {
+                    rangerPolicy = fetchRangerPolicyByResources(atlasRangerService,
+                            "tag",
+                            String.valueOf(provisionalPolicy.getPolicyType()),
+                            provisionalPolicy);
 
-                for (RangerPolicy matchedRangerPolicy : rangerPolicies) {
-                    String labelMatchedPolicyResourcesSignature = new RangerPolicyResourceSignature(matchedRangerPolicy).getSignature();
+                } else if (rangerPolicies.size() == 1) {
+                    rangerPolicy = rangerPolicies.get(0);
+                } else {
+                    String resourcesSignature = new RangerPolicyResourceSignature(provisionalPolicy).getSignature();
 
-                    if (resourcesSignature.equals(labelMatchedPolicyResourcesSignature)) {
-                        rangerPolicy = matchedRangerPolicy;
+                    for (RangerPolicy matchedRangerPolicy : rangerPolicies) {
+                        String labelMatchedPolicyResourcesSignature = new RangerPolicyResourceSignature(matchedRangerPolicy).getSignature();
+
+                        if (resourcesSignature.equals(labelMatchedPolicyResourcesSignature)) {
+                            rangerPolicy = matchedRangerPolicy;
+                        }
                     }
                 }
-            }
 
-            if (rangerPolicy == null) {
-                atlasRangerService.createRangerPolicy(provisionalPolicy);
-            } else {
-
-                rangerPolicy.setPolicyItems(provisionalPolicy.getPolicyItems());
-                rangerPolicy.setDenyPolicyItems(provisionalPolicy.getDenyPolicyItems());
-                rangerPolicy.setDataMaskPolicyItems(provisionalPolicy.getDataMaskPolicyItems());
-
-                if (context.isDeletePurposePolicy()) {
-                    rangerPolicy.getPolicyLabels().remove(getPurposePolicyLabel(context.getPurposePolicy().getGuid()));
+                if (rangerPolicy == null) {
+                    ret = atlasRangerService.createRangerPolicy(provisionalPolicy);
                 } else {
-                    rangerPolicy.getPolicyLabels().addAll(provisionalPolicy.getPolicyLabels());
-                }
 
-                atlasRangerService.updateRangerPolicy(rangerPolicy);
+                    rangerPolicy.setPolicyItems(provisionalPolicy.getPolicyItems());
+                    rangerPolicy.setDenyPolicyItems(provisionalPolicy.getDenyPolicyItems());
+                    rangerPolicy.setDataMaskPolicyItems(provisionalPolicy.getDataMaskPolicyItems());
+
+                    if (context.isDeletePurposePolicy()) {
+                        rangerPolicy.getPolicyLabels().remove(getPurposePolicyLabel(context.getPurposePolicy().getGuid()));
+                    } else {
+                        rangerPolicy.getPolicyLabels().addAll(provisionalPolicy.getPolicyLabels());
+                    }
+
+                    ret = atlasRangerService.updateRangerPolicy(rangerPolicy);
+
+                    LOG.info("Updated Ranger Policy with ID {}", ret.getId());
+                }
             }
+        } finally {
+            RequestContext.get().endMetricRecord(metricRecorder);
         }
 
         RequestContext.get().endMetricRecord(metricRecorder);
@@ -383,164 +399,6 @@ public class AtlasPurposeService {
         if (accessPolicy != null) {
             accessPolicy.setResources(resources);
             atlasRangerService.updateRangerPolicy(accessPolicy);
-        }
-    }
-
-    private List<RangerPolicy> purposePolicyToRangerPolicy(PurposeContext context) throws AtlasBaseException {
-        AtlasPerfMetrics.MetricRecorder metricRecorder = RequestContext.get().startMetricRecord("purposePolicyToRangerPolicy");
-
-        List<RangerPolicy> ret = new ArrayList<>();
-        RangerPolicy rangerAccessPolicy, rangerMaskPolicy;
-
-        AtlasEntity purpose = context.getPurpose();
-
-        List<String> tags = getTags(purpose);
-        if (CollectionUtils.isEmpty(tags)) {
-            throw new AtlasBaseException("Tags list is empty");
-        }
-
-        rangerAccessPolicy = getRangerPolicy(context, 0);
-        rangerMaskPolicy = getRangerPolicy(context, 1);
-
-        Map<String, RangerPolicy.RangerPolicyResource> resources = new HashMap<>();
-        resources.put(RESOURCE_TAG, new RangerPolicy.RangerPolicyResource(tags, false, false));
-
-        rangerAccessPolicy.setResources(resources);
-        rangerMaskPolicy.setResources(resources);
-
-        policyToRangerPolicy(context, rangerAccessPolicy, rangerMaskPolicy);
-
-        if (CollectionUtils.isNotEmpty(rangerAccessPolicy.getPolicyItems()) || CollectionUtils.isNotEmpty(rangerAccessPolicy.getDenyPolicyItems())) {
-            ret.add(rangerAccessPolicy);
-        }
-
-        if (CollectionUtils.isNotEmpty(rangerMaskPolicy.getDataMaskPolicyItems())) {
-            ret.add(rangerMaskPolicy);
-        }
-
-        RequestContext.get().endMetricRecord(metricRecorder);
-        return ret;
-    }
-
-    private void policyToRangerPolicy(PurposeContext context, RangerPolicy rangerAccessPolicy, RangerPolicy rangerMaskPolicy) {
-        rangerAccessPolicy.setName("purpose-tag-" + UUID.randomUUID());
-
-        for (AtlasEntity policy : getPolicies(context.getPurposeExtInfo())) {
-            if (context.isDeletePurposePolicy() && policy.getGuid().equals(context.getPurposePolicy().getGuid())) {
-                continue;
-            }
-
-            List<String> actions = getActions(policy);
-            List<RangerPolicyItemAccess> accesses = new ArrayList<>();
-
-            if (isDataPolicy(policy)) {
-                for (String action : actions) {
-                    accesses.add(new RangerPolicyItemAccess(formatAccessType("heka", action)));
-                }
-
-                if (isDataMaskPolicy(policy)) {
-                    rangerMaskPolicy.setName("purpose-tag-mask-" + UUID.randomUUID());
-                    String maskType = getDataPolicyMaskType(policy);
-                    RangerPolicyItemDataMaskInfo maskInfo = new RangerPolicyItemDataMaskInfo(formatMaskType(maskType), null, null);
-                    setPolicyItem(policy, rangerMaskPolicy, accesses, maskInfo);
-                    rangerMaskPolicy.getPolicyLabels().add(getPurposePolicyLabel(policy.getGuid()));
-
-                } else {
-                    setPolicyItem(policy, rangerAccessPolicy, accesses, null);
-                    rangerAccessPolicy.getPolicyLabels().add(getPurposePolicyLabel(policy.getGuid()));
-                }
-            } else {
-                for (String action : actions) {
-                    if (action.equals(LINK_ASSET_ACTION)) {
-                        accesses.add(new RangerPolicyItemAccess(formatAccessType(ACCESS_ADD_REL)));
-                        accesses.add(new RangerPolicyItemAccess(formatAccessType(ACCESS_UPDATE_REL)));
-                        accesses.add(new RangerPolicyItemAccess(formatAccessType(ACCESS_REMOVE_REL)));
-                    } else {
-                        accesses.add(new RangerPolicyItemAccess(formatAccessType(action)));
-                    }
-                }
-                setPolicyItem(policy, rangerAccessPolicy, accesses, null);
-                rangerAccessPolicy.getPolicyLabels().add(getPurposePolicyLabel(policy.getGuid()));
-            }
-        }
-    }
-
-    private void setPolicyItem(AtlasEntity purposePolicy, RangerPolicy rangerPolicy,
-                               List<RangerPolicyItemAccess> accesses, RangerPolicyItemDataMaskInfo maskInfo) {
-
-        List<String> users = getPolicyUsers(purposePolicy);
-        List<String> groups = getPolicyGroups(purposePolicy);
-        if (getIsAllUsers(purposePolicy)) {
-            users = null;
-            groups = Collections.singletonList("public");
-        }
-
-        if (maskInfo != null) {
-            RangerDataMaskPolicyItem policyItem = new RangerDataMaskPolicyItem(accesses, maskInfo, users, groups, null, null, false);
-            rangerPolicy.getDataMaskPolicyItems().add(policyItem);
-
-        } else {
-            RangerPolicy.RangerPolicyItem policyItem = new RangerPolicy.RangerPolicyItem(accesses, users, groups, null, null, false);
-
-            if (getIsAllow(purposePolicy)) {
-                rangerPolicy.getPolicyItems().addAll(Collections.singletonList(policyItem));
-            } else {
-                rangerPolicy.getDenyPolicyItems().addAll(Collections.singletonList(policyItem));
-            }
-        }
-    }
-
-    private RangerPolicy getRangerPolicy(PurposeContext context, int policyType){
-        RangerPolicy rangerPolicy = new RangerPolicy();
-        AtlasEntity purpose = context.getPurpose();
-
-        rangerPolicy.setPolicyLabels(Arrays.asList(getPurposeLabel(purpose.getGuid()), "type:purpose"));
-
-        rangerPolicy.setPolicyType(policyType);
-        rangerPolicy.setServiceType("tag");
-        rangerPolicy.setService("default_atlan");
-
-        return rangerPolicy;
-    }
-
-    private void verifyUniqueNameForPurposePolicy(PurposeContext context) throws AtlasBaseException {
-
-        if (!context.isCreateNewPurposePolicy() && !getName(context.getExistingPurposePolicy()).equals(getName(context.getPurpose()))) {
-            return;
-        }
-        List<String> policyNames = new ArrayList<>();
-        String newPolicyName = getName(context.getPurposePolicy());
-
-        List<AtlasEntity> policies = getPolicies(context.getPurposeExtInfo());
-
-        if (CollectionUtils.isNotEmpty(policies)) {
-            if (context.isCreateNewPurposePolicy()) {
-                policies = policies.stream()
-                        .filter(x -> !x.getGuid().equals(context.getPurposePolicy().getGuid()))
-                        .collect(Collectors.toList());
-            }
-        }
-
-        policies.forEach(x -> policyNames.add(getName(x)));
-
-        if (policyNames.contains(newPolicyName)) {
-            throw new AtlasBaseException(PURPOSE_ALREADY_EXISTS, newPolicyName);
-        }
-    }
-
-    private void validatePurposePolicyRequest(PurposeContext context) throws AtlasBaseException {
-        if (!AtlasEntity.Status.ACTIVE.equals(context.getPurpose().getStatus())) {
-            throw new AtlasBaseException(OPERATION_NOT_SUPPORTED, "Purpose not Active");
-        }
-
-        if (CollectionUtils.isEmpty(getActions(context.getPurposePolicy()))) {
-            throw new AtlasBaseException(BAD_REQUEST, "Please provide actions for policy policy");
-        }
-
-        if (CollectionUtils.isEmpty(getPolicyUsers(context.getPurposePolicy())) &&
-                CollectionUtils.isEmpty(getPolicyGroups(context.getPurposePolicy())) && !getIsAllUsers(context.getPurposePolicy())) {
-
-            throw new AtlasBaseException(BAD_REQUEST, "Please provide users/groups OR select All users for policy policy");
         }
     }
 }
